requirements.txt
Django==1.4.2
djangorestframework==2.1.6
wsgiref==0.1.2
